removed main controller and organized the rest apis
